rename nipoppy.config.base -> nipoppy.config.main
