+++ nipoppy_cli/nipoppy/config/__init__.py
@@ -1,6 +1,6 @@
 """User/pipeline configurations."""
 
-from .base import Config
 from .boutiques import BoutiquesConfig
+from .main import Config
 from .pipeline import PipelineConfig
 from .singularity import SingularityConfig
